cart finished. If not authenticated order shows no order Fixes we can delete users who don`t have a cart

+++ src/main/java/com/example/demo/service/UsersService.java
@@ -166,7 +166,10 @@ public void deleteUser(Long userId) {
         }
 
         // Удалить Cart
-        cartRepository.delete(user.getCart());
+        if (user.getCart() != null){
+            cartRepository.delete(user.getCart());
+        }
+
 
         // Удалить Users
         usersRepository.delete(user);
